Drop web seeds from torrent magnet URIs
They've been pointing to the 'old' URL used before the website upgrade,
so it will return 404.

In my opinion, web seeds defeats the point of BitTorrent as the 'direct'
HTTP download will always be available as an alternate. Plus, a web seed
will eventually break when the file gets moved into /archived/.