Fix RSS modification dates
Some RSS clients use <updated> instead of <published> to determine
changes. Prior to this, the dates reset to the time of deployment.

This corrects the dates by using the git timestamp for the file. This
does NOT fix RSS localized versions of the site, as they are generated
at build time.