Remove external WebP generation option
These days, browsers natively understand WebP, and offers superior
compression to JPG/PNG in many cases. Prefer to use WebP for the source,
rather then have two copies which may not be used anymore (under
Cloudflare Pages)

Previously, the nginx server had a rule that serves the WebP in place of
the PNG/JPG, as a bandwidth optimisation approach.