feat: update downloads for 24.04 from 24.04.1 to 24.04.2 (#183)
* feat: update downloads for 24.04 from 24.04.1 to 24.04.2

According to:

https://lists.ubuntu.com/archives/ubuntu-announce/2025-February/000308.html

... Ubuntu 24.04.2 LTS (2nd point release of "Noble Numbat") was released today (20th February 2025) 
for Ubuntu and its several flavors / flavours (including "Ubuntu MATE").

The download links at:

https://ubuntu-mate.org/download/amd64/noble/

... are currently still pointing to :

https://cdimage.ubuntu.com/ubuntu-mate/releases/noble/release/ubuntu-mate-24.04.1-desktop-amd64.iso


This commit and Pull Request replaces "24.04.1" by "24.04.2" in the version number and updates the location (URL) and SHA256 checksum of the ISO file and updates the Magnet link accordingly.

* Update downloads.yml : replace "24.04.1 LTS" by "24.04.2 LTS" in "name:"

In my previous commit, I forgot to also replace "24.04.1 LTS" by "24.04.2 LTS" in the "name:" of the "noble:" block. 

This additional commit makes that change.